+++ aql-engine/src/main/java/org/ehrbase/openehr/aqlengine/sql/EncapsulatingQueryUtils.java
@@ -46,6 +46,7 @@ import org.ehrbase.openehr.aqlengine.asl.model.field.AslComplexExtractedColumnFi
 import org.ehrbase.openehr.aqlengine.asl.model.field.AslConstantField;
 import org.ehrbase.openehr.aqlengine.asl.model.field.AslDvOrderedColumnField;
 import org.ehrbase.openehr.aqlengine.asl.model.field.AslField;
+import org.ehrbase.openehr.aqlengine.asl.model.field.AslFolderItemIdVirtualField;
 import org.ehrbase.openehr.aqlengine.asl.model.field.AslOrderByField;
 import org.ehrbase.openehr.aqlengine.asl.model.field.AslSubqueryField;
 import org.ehrbase.openehr.aqlengine.asl.model.join.AslJoin;
@@ -107,7 +108,6 @@ final class EncapsulatingQueryUtils {
         return switch (af.getBaseField()) {
             case null -> null;
             case AslColumnField f -> FieldUtils.field(Objects.requireNonNull(src), f, true);
-            case AslAggregatingField __ -> throw new IllegalArgumentException("Can't aggregate on AslAggregatingField");
             case AslComplexExtractedColumnField ecf -> {
                 Objects.requireNonNull(src);
                 yield switch (ecf.getExtractedColumn()) {
@@ -139,6 +139,10 @@ final class EncapsulatingQueryUtils {
             }
             case AslConstantField cf -> DSL.inline(cf.getValue(), cf.getType());
             case AslSubqueryField sqfd -> subqueryField(sqfd, aslQueryToTable);
+            case AslAggregatingField __ -> throw new IllegalArgumentException(
+                    "Cannot aggregate on AslAggregatingField");
+            case AslFolderItemIdVirtualField __ -> throw new IllegalArgumentException(
+                    "Cannot aggregate on AslFolderItemIdValuesColumnField");
         };
     }
 
@@ -296,6 +300,8 @@ final class EncapsulatingQueryUtils {
             case AslAggregatingField af -> sqlAggregatingField(af, src, aslQueryToTable);
             case AslConstantField<?> cf -> DSL.inline(cf.getValue(), cf.getType());
             case AslSubqueryField sqf -> subqueryField(sqf, aslQueryToTable);
+            case AslFolderItemIdVirtualField fidv -> throw new IllegalArgumentException(
+                    "%s is not support as select field".formatted(fidv.getExtractedColumn()));
         };
     }
 
@@ -320,10 +326,12 @@ final class EncapsulatingQueryUtils {
                             "%s is not a complex extracted column".formatted(ecf.getExtractedColumn()));
                 }
             }
-            case AslAggregatingField __ -> throw new IllegalArgumentException(
-                    "Cannot aggregate by AslAggregatingField");
             case AslSubqueryField sqf -> Stream.of(subqueryField(sqf, aslQueryToTable));
             case AslConstantField __ -> Stream.empty();
+            case AslAggregatingField __ -> throw new IllegalArgumentException(
+                    "Cannot aggregate by AslAggregatingField");
+            case AslFolderItemIdVirtualField __ -> throw new IllegalArgumentException(
+                    "Cannot aggregate by AslFolderItemIdValuesColumnField");
         };
     }
 
@@ -351,10 +359,12 @@ final class EncapsulatingQueryUtils {
                             (Field<JSONB>) FieldUtils.field(src, f, true)));
                     case AslColumnField f -> columnOrderField(f, src, templateService);
                     case AslComplexExtractedColumnField ecf -> complexExtractedColumnOrderByFields(ecf, src);
-                    case AslAggregatingField __ -> throw new IllegalArgumentException(
-                            "ORDER BY AslAggregatingField is not allowed");
                     case AslConstantField __ -> Stream.<Field<?>>empty();
                     case AslSubqueryField sqf -> Stream.of(subqueryField(sqf, aslQueryToTable));
+                    case AslAggregatingField __ -> throw new IllegalArgumentException(
+                            "ORDER BY AslAggregatingField is not allowed");
+                    case AslFolderItemIdVirtualField __ -> throw new IllegalArgumentException(
+                            "ORDER BY AslFolderItemIdValuesColumnField is not allowed");
                 })
                 .map(f -> f.sort(ob.direction()));
     }