+++ aql-engine/src/test/java/org/ehrbase/openehr/aqlengine/asl/AslGraph.java
@@ -45,10 +45,12 @@ import org.ehrbase.openehr.aqlengine.asl.model.field.AslColumnField;
 import org.ehrbase.openehr.aqlengine.asl.model.field.AslComplexExtractedColumnField;
 import org.ehrbase.openehr.aqlengine.asl.model.field.AslConstantField;
 import org.ehrbase.openehr.aqlengine.asl.model.field.AslField;
+import org.ehrbase.openehr.aqlengine.asl.model.field.AslFolderItemIdVirtualField;
 import org.ehrbase.openehr.aqlengine.asl.model.field.AslOrderByField;
 import org.ehrbase.openehr.aqlengine.asl.model.field.AslSubqueryField;
 import org.ehrbase.openehr.aqlengine.asl.model.join.AslAuditDetailsJoinCondition;
 import org.ehrbase.openehr.aqlengine.asl.model.join.AslDelegatingJoinCondition;
+import org.ehrbase.openehr.aqlengine.asl.model.join.AslFolderItemJoinCondition;
 import org.ehrbase.openehr.aqlengine.asl.model.join.AslJoin;
 import org.ehrbase.openehr.aqlengine.asl.model.join.AslJoinCondition;
 import org.ehrbase.openehr.aqlengine.asl.model.join.AslPathFilterJoinCondition;
@@ -224,6 +226,12 @@ public class AslGraph {
                             .formatted(
                                     c.getLeftOwner().getAlias(),
                                     c.getRightOwner().getAlias());
+                    case AslFolderItemJoinCondition
+                    c -> "FolderItemJoinCondition FOLDER -> %s [%s.vo_id in %s.data.items[].id.value]"
+                            .formatted(
+                                    c.descendantRelation(),
+                                    c.getRightOwner().getAlias(),
+                                    c.getLeftOwner().getAlias());
                 })
                 .map(s -> indented(level, s))
                 .collect(Collectors.joining());
@@ -266,6 +274,7 @@ public class AslGraph {
                                             .map(c -> conditionToGraph(level + 2, c))
                                             .collect(Collectors.joining("\n", "", "")));
             case AslConstantField f -> "CONSTANT (%s): %s".formatted(f.getType().getSimpleName(), f.getValue());
+            case AslFolderItemIdVirtualField f -> providerAlias + f.aliasedName() + " -- FOLDER.items";
         };
     }